requirements.txt
pytest==6.2.5
pytest-cov==4.0.0
python-dotenv==0.21.0
boto3==1.24.88